refactor: Pool creation example combined with init. Made single example.
